server/requirements.txt
Flask
flasgger
pydub
flask-sock
rpi-lgpio